rdp/rdpgfx: Target render surface, when encoding data with AVC
A surface may specify another surface to render the frame content to.
The intention for this is to handle situations, where AVC encoder
requirements don't match the RDPGFX protocol requirements.
In such case, instruct the client to render the frame content to the
separate offscreen surface, that uses an aligned size, that matches the
encoder requirements and then instruct the client to blit the actual
affected area from the separate offscreen surface onto the mapped
surface using SURFACE_TO_SURFACE PDUs.
These operations are handled within a logical frame. This means, that
the client MUST handle all operations of the logical frame, before
displaying the frame and submitting the frame ack.