rdp-graphics-pipeline: Create separate render surface when needed
The documentation of the graphics pipeline says:
[...] "the width and height of the MPEG-4 AVC/H.264 codec bitstream
MUST be aligned to a multiple of 16 and MUST be cropped by the region
mask specified in the regionRects field that is embedded in the
avc420MetaData field." ([MS-RDPEGFX] 2.2.4.4 RFX_AVC420_BITMAP_STREAM)

For the NVENC session, the height of each frame needs to be aligned to
a multiple of 64.
64 is a multiple of 16, but mstsc actually requires the bitstream to be
aligned to 16x16 tiles, as long as these tiles cover the surface.
This last part is not mentioned in the documentation, but this applies
to a lot of things of the RDP protocol.
In case of an error in the documentation, the behaviour of mstsc and
the RDS in MS Windows determine the behaviour of the RDP protocol.
So, adjust the implementation in gnome-remote-desktop accordingly.

Whenever the aligned height of the NVENC session, which is a multiple
of 64, does not match the aligned height of the mapped surface, create
a separate render surface.
This render surface will then have the aligned size of the NVENC
session and will be used to render the data to on the client side.
With the work in the previous commits, the client is instructed to blit
from the render surface to the mapped surface, when the render surface
needs to be used.

Without these changes, mstsc would abort the remote desktop connection
upon receiving a RDPGFX_WIRE_TO_SURFACE_PDU_1 message with a protocol
error.