vnc-server: Cleanup resources directly when stopping the server
To stop the VNC server, gnome-remote-desktop currently calls
grd_vnc_server_stop() to stop the socket listener and then unrefs the
server instance, where also all current VNC sessions are supposed to be
cleaned up.
This works fine, when just the backend is disabled, but
gnome-remote-desktop continues to run.
However, when the daemon is shut down, this is not the case:
Although the server backends are cleaned up before the GrdContext
instance, the context instance is in reality cleaned up first.
This is visible in the form that glib does not run the dispose()
function in the server class directly.
The reason for this is that the GSource for the socket callback also
has its internal reference.
This GSource cleans up its reference, when the G_IO_NVAL value is
returned from its fd.

The effect of the current handling is that gnome-remote-desktop
crashes, when the daemon exits.
To prevent this crash, clean up the internal server resources in
grd_vnc_server_stop() too.
This ensures that the destruction order is kept and no use-after-free
situations happen, when the daemon exits.