hwaccel-nvidia: Unload modules before releasing CUDA context
The lifetime of CUDA modules depends on the lifetime of the current
CUDA context.
This information is hidden and apparently only documented in the
function "cuFuncGetModule".
Unloading CUDA modules after releasing the CUDA context can sometimes,
but not always lead to a crash.

So, fix this situation by unloading the CUDA modules before releasing
the CUDA context.