daemon: Handle SIGINT and SIGTERM signals properly
Currently, when the gnome-remote-desktop-daemon is stopped via a SIGINT
or SIGTERM signal, it directly stops without cleaning up its resources.
This is especially visible, when there is a current RDP session
running, as all FUSE mounts will remain mounted, when
gnome-remote-desktop is terminated via these signals.
This is very annoying to see, when e.g. running the command `df -h`, as
it will show that the socket for the FUSE mount is not connected any
more.
The SIGINT signal is emitted, when running the daemon manually in a
terminal window, the SIGTERM signal is emitted by systemd, when
stopping or restarting the daemon.
systemd handles SIGTERM signals gracefully, if the service handles
these signals.
This means systemd will wait until gnome-remote-desktop stopped all
sessions and cleaned up its resources.
This also includes the FUSE unmount operations.

To handle the SIGINT and SIGTERM signals properly, register handlers
for them, which shut down the daemon, when the respective signals are
received.