daemon: Only start servers, when they are not started yet
Triggering a start of a server backend can be done in two ways:
1. In maybe_enable_services(), when the dbus proxies are ready
2. When switching the 'enable' gsettings key of the respected backend

If both events are triggered at the same time, then a start is
triggered in `maybe_enable_services()` and, depending on the backend,
in `on_rdp_enabled_changed()` or in `on_vnc_enabled_changed()`.
This leads to an assertion hit, as the respective server was already
started in such case.

To fix this situation, change the assertion into a normal conditional
statement, where the respective server backend is only started, when it
is not already started.

Fixes: https://errors.ubuntu.com/problem/11540483956116d59b3da5908cb6f7b0caeedae4