rdp: Retrieve frame members before stealing frame pointer
When a frame only contains a pointer update, but no frame data, the
pointer data is directly copied and the on_frame_ready() callback is
directly executed.
The callback and the callback user data are part of the frame, and the
frame is an auto pointer.
As a result, g_steal_pointer() needs to be used to not free the frame,
when the pipewire buffers were chosen.
However, g_steal_pointer() may be executed first, and in such case
dereferencing the frame pointer to gain access to its members will fail
and lead to a crash.

To fix this issue, copy the pointer values of the members first, before
stealing the pointer value of the frame.

Fixes: https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/issues/96