context: Initialize EGL thread when daemon is ready
Since recently, gnome-remote-desktops systemd service file requires the
WantedBy property in order to be started automatically upon login.
However, it does not seem to be possible to defer the launch until
gnome-shell is also started.
As a result, the EGL thread never initializes with the platform
EGL_PLATFORM_WAYLAND_EXT.
gnome-remote-desktop, however, does know when gnome-shell is started,
as it waits on mutters remote desktop and screencast dbus interfaces to
appear, before starting the server backends.

Do the same for the EGL thread by adding a new method to notify the
context, when the daemon is ready.
Hook up to this method in maybe_enable_services() to ensure, that the
EGL thread can be initialized with EGL_PLATFORM_WAYLAND_EXT.

Closes: https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/issues/89