hwaccel-nvidia: Degrade a warning to a normal message
In multi-GPU setups, the EGL thread will initialize on the primary GPU.
In that case, CUDA will initialize, but will not find the NVidia GPU
via the EGL thread, as it is not the primary one.

This situation is not an error and it is already correctly handled.
So, just degrade the warning to a normal message.