tests/egl: Replace tautologies with actual modifier checks
The DRM format modifier should only be used, when it is not invalid.
The current checks currently do not correctly check for this.
Instead, the existence of the modifiers array in the DmaBuf struct is
always checked, which is always true.

So, fix this by replacing this tautology with an actual modifier check.