rdp/nw-auto: Fix initial ping source not being created
Currently, gnome-remote-desktop does not create the initial GSource to
emit pings, unless ping interval changes to low due to inactivity on
the screen.
As a result, stutters can appear.

To fix this issue, introduce a new ping internal with the name "NONE".
This new interval represents the absence of the ping source.
When the client emits the SuppressOutput PDU, where
gnome-remote-desktop stops emitting pings, also reset the ping
interval.