rdp/nw-auto: Also notify client about network characteristics results
RDP clients can tell the user about the quality of the network
connection.
This is done with the help of the network characteristics results, that
are sent by the server.
The RDP protocol currently allows these characteristics to be sent in
the following combinations:

1. Base RTT, and average RTT value
2. Bandwidth speed, and average RTT value
3. Base RTT, bandwidth speed, and average RTT value

FreeRDP currently implements combination 1 and 3.
So, to be able to send these characteristics, also determine the base
round trip time value.
Then send the Network Characteristics Result PDU, when receiving RTT
responses.
Since, for the server side only RTT values with ping intervals way
lower than one second are useful, Network Characteristics Result PDUs
should not be sent upon retrieval of any RTT response, as otherwise the
RDP client might update the network indicator very rapidly.
Usually, ping values in applications are updated every second, so add
additional handling to ensure the same behaviour is achieved.