rdp/nw-auto: Add APIs for bandwidth measurement
mstsc has a network indicator icon, which is only shown, when the
server side also sends Network Characteristics Result PDUs containing a
bandwidth measurement result.
Currently, gnome-remote-desktop does not measure the bandwidth, so add
additional APIs to be able to perform these measurements.
Measuring the bandwidth is useful for screen updates.
With the graphics pipeline, progressive rendering is possible, which is
helpful for low bandwidth scenarios.
Currently, it is not yet implemented, but will hopefully be in the
future.
Since sending screen updates in the graphics pipeline is done in an
async way, add an additional API to be able to queue a bandwidth
measure stop.
To be able to be notified of such queued bandwidth measure stop, add an
API to retrieve the associated WinPR event handle.