rdp: Add API to tear down channels during a session
When during a session a channel is not needed any more, it should not
waste resources.
This will especially be the case for the AUDIO_PLAYBACK_DVC channel, as
the user might disable that channel on the client side.
Since opening a dynamic channel may not imply, that the client supports
that channel, add an API to tear down channels during a session.

This API will be used in the next commits for the AUDIO_PLAYBACK_DVC
channel.