rdp: Add channel for client metrics notifications
The telemetry channel is a channel with just one PDU. It is sent by
remote desktop clients supporting the channel, after the first graphics
update is received via the graphics pipeline.
It is useful, when trying to minimize the time until a remote desktop
session is established.
Upon receiving the telemetry PDU, log the received client metrics.