session-rdp: Add API to subscribe to DVC creation statuses
Opening a dynamic virtual channel (DVC) is not an instant action.
Instead, the server side sends a DVC Create Request PDU
(DYNVC_CREATE_REQ) and the client side answer with a DVC Create
Response PDU (DYNVC_CREATE_RSP).
Only if the creation status value is non-negative, the creation was
successful.
With the latest FreeRDP stable update, it is now possible to get
notified of that creation status.

So, use that API in FreeRDP to add an API in gnome-remote-desktop to
allow DVCs to subscribe to their creation status.
This API will later be used by DVCs implemented in
gnome-remote-desktop.