rdp/nw-auto: Turn assertion into normal conditional statement
For unknown reasons, the initial state of the bandwidth measure stop
event is set to TRUE on some systems, although the opposite should be
the case, leading to an assertion hit.

So, turn for now the assertion into a normal condition.
In addition to that, always reset the event before hitting the new
condition to avoid further unnecessary wakeups.

Closes: https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/issues/102