rdp: Add classes for camera redirection
Camera redirection in RDP is a feature that allows using cameras that
physically exist on the client side to be used inside a remote desktop
session on the server side.
The way camera redirection works in RDP is by using dynamic virtual
channels (DVCs). First, there is the camera device enumerator channel:
In this channel, the protocol version is negotiated and clients announce
added and removed cameras here. In such an announcement, the client
tells the server side the name of the camera device to be used and the
DVC name where the actual camera handling happens, as each camera device
uses its own DVC. The server side can then open that DVC, and initialize
the actual handling.

When initializing a camera device, the server side first needs to know
what kind of streams are supported by the camera, and what media types
are used in each stream. For this, use a state machine in order to
ensure an ordered initialization. After all required data is received,
the camera streams themselves can be created. These are the usual
PipeWire streams that just forward any frame request as a sample request
to the client. When such sample is received, it is forwarded to a
decoder, which then decodes the frame and notifies the stream about the
result.

Unfortunately, camera streams can in [MS-RDPECAM] not be individually
stopped. However, that is necessary when a camera stream itself is
stopped or when the stream format changes.
To handle this issue, keep track of all stream states. If one camera
stream needs to be stopped, stop all streams, and restart all previously
running ones again that were not stopped. In practice, this will though
rarely happen, as most camera devices only provide one stream in RDP.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/360>