Add TPM 2.0 helper class
This is intended to be used to implement credentials storage better than
plain text on the file system by using the Trusted Platform Module
(TPM).

It consists of creating a high level TPM helper object that is used
either to store or to read the secret. Testing shows that using the same
session/TPM transmission instance ends up with memory constraint issues,
so the usage is split into two completely sepearate sessions, avoiding
that issue.

Storing the secret uses the Platform Configuration Register (PCR) to
create an authorization policy, which is used to encrypt the secret
using the TPM module.

The same PCRs are then used to create an equivalent authorization used
when using the TPM module te decrypt the secret.

The encrypted secret itself is stored as a GVariant containing the
TPMS_CONTEXT properties.

Test cases are added to test storing and restoring the secret. Since one
can only restore what was stored on the same machine, the "restore" test
relies on the "store" test to run first, and will otherwise fail.

Guard the tss2_*.h include macros in attributes ignoring deprecated
declaration warnings. This is needed for certain versions of tpm2-tss;
see https://github.com/tpm2-software/tpm2-tss/issues/2371 for details.