Introduce GrdCredentials
This type aims to provide an abstraction level above credentials
management, so that they can be handled differently depending on in what
way the daemon is run.

For now, the only added backend is one using libsecret, and it is
identical to the existing libsecret implementation. Every user that used
libsecret directly now access the credentials via the abstracted API.

This changes the stored format when using the VNC backend. To handle
this, change the Secret schema name, and migrate from the old format
whenever the Secret credentials instance is created, i.e. either before
initializing the daemon, or before running any grdctl command.