session-rdp: Stop leaking session sources when session creation fails
Currently, the monitor layout update source is leaked, when the session
creation fails.
This GSource, however, must be destroyed, when the session is stopped
and not later.

To handle both the success and fail case, add a helper method, where
the existing session sources are cleared, when the session stops or
when it is disposed.