rdp: Initialize graphics pipeline when the session is ready
Currently, the graphics pipeline is created in the socket thread after
the capability exchange, while other virtual channels are currently
created when the actual remote desktop session is ready, as these
channels cannot be created before or need to be created in the main
thread.
For consistency reasons do the same for the graphics pipeline.
This also has the advantage, that the graphics pipeline does not need
to be shut down immediately after creation, when starting the remote
desktop session fails.

Since the telemetry channel depends on the graphics pipeline, apply
this change also for that virtual channel.