rdp-pipewire-stream: Skip corrupted buffers
With [0] mutter now also sends additional header data for each SPA
buffer.
This header contains a flag field, where mutter can specify, that a SPA
buffer is corrupted.
This flag is necessary, as queued PipeWire buffers, which wrap around
SPA buffers, MUST be queued again, after dequeuing them.
So, if setting the data in the SPA buffer fails, the buffer MUST be
ignored.

So, look for the header data in each SPA buffer. If available, and the
SPA_META_HEADER_FLAG_CORRUPTED flag is set, skip the buffer, instead of
trying to process it.

[0]: https://gitlab.gnome.org/GNOME/mutter/-/commit/30b5229e0eeb5117fe3061cfcc1d1e59c199f4d9