build: Stop linking the server against libfreerdp-client2
The clipboard implementation in the RDP backend requires a function in
FreeRDP to parse received file lists from clients.
This function was previously only available in the client side library
in FreeRDP.
With the FreeRDP 2.8.0 release, the function definition is now
available in libfreerdp2 too, so remove the libfreerdp-client2
requirement.