rdp/nw-auto: Disallow unqueued bandwidth measure stops
Recently, it was discovered, that WinPR events are handled in
unexpected manners on a few systems.
In this case here, a Bandwidth Measure Stop PDU can be sent before it
is meant to be sent.
This results into an unexpected state of the bandwidth measurement
state machine, as the bandwidth measure stop was meant to be queued
before it is actually stopped.
Since there are currently no places, where a Bandwidth Measure Stop PDU
is sent directly, disallow unqueued bandwidth measure stops.

Fixes: https://errors.ubuntu.com/problem/98a015184483b3876c1a63cfb0191c51ab2df6e5