clipboard-rdp: Get rid of all WaitForSingleObject calls
Getting rid of the WaitForSingleObject() call on the
completed_format_list_event is here trivial, as it is basically the
equivalent action as checking for the source id of the server format
list update.

To get rid of the other WaitForSingleObject() calls, which all check
here the stop_event, introduce a simple boolean value, which is set to
TRUE, when the clipboard instance is torn down.