rdp: Get rid of stop_event handling during initialization of DVCs
Recently, it was discovered, that WinPR events are handled in
unexpected manners on a few systems.
To avoid unexpected behaviour, reduce their usage, wherever it is
possible.

In order to get rid of the stop event during initialization of DVCs,
introduce a boolean value, which indicates, whether the session should
stop.
This boolean is only set once to TRUE in a session and is used in the
socket thread to determine, whether a DVC should be initialized or not.
This behaviour is still necessary, as the graphics pipeline is
essential for the session.
If support for the graphics pipeline is advertised, but actually not
supported by the client, it is an error in the client and the session
must be stopped.

When stopping all remote desktop sessions via the gnome-shell submenu,
the RDP backend needs to set the error info to prevent clients from
doing an autoreconnect.
Previously, the stop_event was used here too to determine, whether a
session was stopped via the gnome-shell submenu or not.
Replace this behaviour by checking now, whether an idle session close
source was created.