clipboard-rdp: Get rid of all WaitForMultipleObjects calls
Replacing WaitForMultipleObjects() calls is not necessarily trivial and
cannot be done with a general scheme.
It depends on how, where, and in which context these calls are used.
In clipboard-rdp, they are all used in the CLIPRDR thread, waiting for
actions to complete on the main thread.

So, in order to replace all WaitForMultipleObjects() calls here, use
for each action its own boolean value.
In addition to that, use the existing protocol_stopped boolean value to
check, whether the clipboard instance is torn down.
Use both variables (protocol_stopped, action) in a typical GCond-GMutex
code fragment to replace each WaitForMultipleObjects() call here.