clipboard-rdp: Fix race in debug message
Lock Clipboard Data PDUs can be sent at any point in time after the
clipboard capabilities and temporary directory have been exchanged in
the clipboard initialization sequence ([MS-RDPECLIP] 2.2.4.1).
With mstsc as RDP client, the PDU is sent before the
FormatListResponse.
So, in the locking stage, it is not yet known, that File Content
Requests are successful.
To avoid blocking in the main and CLIPRDR thread, a simple boolean
value, indicating whether requests are allowed or not, is used to
determine that state.

While extremely unlikely, but still possible, it may happen, that a
debug message is printed to the output, while the associated File
Contents Request is successful.

Fix this annoyance by retrieving the state first and then use the local
copied state to determine, whether that debug message should be shown
or not.