rdp-graphics-pipeline: Print unknown cap set versions to the output
The graphics pipeline is an integral part of the modern remote desktop
in the RDP protocol.
When adding new features, Microsoft adds new capability sets as part of
the capability negotiation.
These new capability sets may not be documented instantly or only on
request, and so it is beneficial to print unknown capability versions
to the output to get notified of them.

So, print any unknown capability set versions to the output in order to
become attentive about new functionality.