tpm: Check for a module before attempting to initialize interface
The TPM2 tools work, but only if a TPM2 device is found. If no TPM2
device is found, they don't just fail, unless the daemon for the TPM2
device discovery is not installed.
Instead, they lead to a freeze in the thread of the program, where they
are called.
While this is an error in the TPM2 tools, people will run into this
issue.

So, prevent this issue by manually checking for a TPM2 device first.
If no TPM2 is found, just error out.