rdp/rdpgfx: Only clear main surfaces directly
When overriding the render surface of a GFX surface, the render surface
is managed by its GFX surface.
This includes the destruction of the GFX surface, which is performed,
when its parental surface is destroyed.

When destroying all surfaces, e.g. because the monitor configuration
changed or due to a protocol reset of the graphics pipeline, then
currently all auxiliary surfaces (this includes the separate render
surfaces) are manually destroyed too.
Since they are already managed by their parental surface, a
segmentation fault can occur, when destroying an auxiliary surface.
In practice, this segfault, for some reason, did not happen yet, but
the error is there.
So, on some systems or situations, the segfault is expected to happen.

Fix this issue by marking all separate render surfaces as auxiliary
surfaces.
When clearing all surfaces, only clear the surfaces directly, that are
not marked as auxiliary surface, i.e. surfaces that are main surfaces.
Any auxiliary surface is destroyed, when its parental surface is
destroyed.