session: Only set clipboard pointer, when initialization was successful
When enabling the clipboard, the clipboard instance uses the return
value to determine, whether requests from mutter are possible or not.
When enabling the clipboard failed, the return value is recognized, but
the clipboard pointer in the session class is still set, leading to an
assertion hit, when a request from mutter is received, but the
clipboard was actually never successfully enabled.

To fix this issue, simply only set the clipboard pointer, when enabling
the clipboard was successful.

https://errors.ubuntu.com/problem/e118a2a1b7e6650ee263f50f52fdab8a5dcd18ee