session-rdp: Fix invalid colour depth, when codecs or GFX is advertised
In RDP, all codecs and the graphics pipeline require the peer to
support the colour depth 32 bits per pixel.
While advertising the support for codecs and the graphics pipeline, the
client can still prefer a specific colour depth, although it is
nonsense doing it.
This preferred colour depth setting exists for ancient servers (Win XP
era), but mstsc in recent MS Windows versions still allows a user to
set this setting.
The supported colour depths and support for the graphics pipeline are
submitted by the client in the Client Core Data, which is read before
the Capabilities() callback is called.
The colour depth used in the session is written by the server side in
the Bitmap Capability Set, which is sent after the Capabilities()
callback is called.

When reading the Client Core Data of a client, FreeRDP overrides the
ColorDepth setting, when the preferred colour depth of the client is
lower than the one on the server side.
While FreeRDP clients always automatically set the preferred colour
depth to 32, when any codec or the graphics pipeline is advertised,
mstsc does not.
This results into mstsc dropping the connection, if the user overrides
the preferred colour depth setting.

To fix this issue, check in the Capabilities() callback, whether any
codec or the graphics pipeline was advertised.
In such case, set the colour depth back to 32. This will result into
writing the correct colour depth value in the Bitmap Capability Set and
mstsc not dropping the connection, when the user selected a different
preferred colour depth in mstsc.
MS Windows RDS does the same, when the client supports the graphics
pipeline.

Fixes: https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/issues/118