rdp/rdpsnd: Don't prepare audio samples when volume data is invalid
An audio output stream can already submit audio samples to the audio
playback instance, while its volume values might not be received yet.
In such case, gnome-remote-desktop currently hits an assertion, that
assumes that the amount of volume values is equal to 1, when the amount
of volume values is lower than 2.

To fix this assertion hit, assume that the volume of the stream is
muted, when the amount of volume values is 0.
As soon as the volume data is available, the audio output stream can
submit the audio samples to the pending frames queue.

Fixes: https://errors.ubuntu.com/problem/060bb232251f482381750964fdc92094bed5fe42