rdp/nw-auto: Fix crash when updating unchanged ping source
Commit 51e94ba8dfaa3d6daf31cb982e40e4e0439acb71 fixed an issue, where
the ping source was attempted to be created, when it already existed,
as gnome-remote-desktop previously assumed, that every update of the
RTT consumers and RTT consumer necessities will always change the ping
source.
Commit 351893fac77dea21a44fc44508ba2eefe1d1af4a accidentally broke that
behaviour by always assuming again, that updating the ping source will
always change the ping source itself.
However, this time, the situation is different. Since the former commit
added an assertion to not create a ping source again, when it already
exists, gnome-remote-desktop crashes, due to invalid behaviour (which
also made this issue discoverable).

Fix this situation again by checking again, whether an update of the
ping source is needed.
Due to the latter commit, the initial ping source should be created, as
that commit introduced another state for the ping interval recognizing
it to be non-existent.

Fixes: https://errors.ubuntu.com/problem/4d8b5efdf8ade96d4019c6b7cb92649f660e9fe9