Add class for writing NAL units
While VAAPI provides the ability to encode frames for H.264 on the GPU,
it unfortunately does not write the NAL units, at least not on Intel
GPUs. The NAL units basically contain the config or settings of the
H.264 stream, for example: which profile or QP value was used.
Without knowing this information, the decoder won't know how to treat
the received bitstream.

To address this issue, add a class, that writes the NAL units for a
given H.264 config. Some values are only needed once (per IDR frame),
like the SPS parameter, but others are needed for every frame.
The individual functions, writing the NAL bitstream use as reference the
official H.264 documentation (Rec. ITU-T H.264 (08/2021)). The used
sections are specified in the comments.